refactor: create dedicated navigation options functions
- Create useTabOptions() hook for standardized tab configuration
- Create createModalScreenOptions() function for modal screens
- Refactor tab layout to use new useTabOptions helper
- Update imports to use new navigation structure
- Convert index.ts files to index.tsx for consistency
- Add tab badge example to explore screen
