+++ src/app/(tabs)/_layout.tsx
@@ -1,31 +1,14 @@
 import { Tabs } from 'expo-router';
 import React from 'react';
-import { Platform } from 'react-native';
 
-import { HapticTab } from '@/components/haptic-tab';
 import { IconSymbol } from '@/components/ui/icon-symbol';
-import TabBarBackground from '@/components/ui/tab-bar-background';
-import { useThemeColor } from '@/hooks/use-theme-color';
+import { useTabOptions } from '@/components/layout/navigation';
 
 export default function TabLayout() {
-  const tintColor = useThemeColor('tint');
+  const tabOptions = useTabOptions();
 
   return (
-    <Tabs
-      screenOptions={{
-        tabBarActiveTintColor: tintColor,
-        headerShown: false,
-        tabBarButton: HapticTab,
-        tabBarBackground: TabBarBackground,
-        tabBarStyle: Platform.select({
-          ios: {
-            // Use a transparent background on iOS to show the blur effect
-            position: 'absolute',
-          },
-          default: {},
-        }),
-      }}
-    >
+    <Tabs screenOptions={tabOptions}>
       <Tabs.Screen
         name="index"
         options={{
@@ -38,6 +21,7 @@ export default function TabLayout() {
         options={{
           title: 'Explore',
           tabBarIcon: ({ color }) => <IconSymbol size={28} name="paperplane.fill" color={color} />,
+          tabBarBadge: 5,
         }}
       />
     </Tabs>
